# backend/api/public-openapi.yaml
...
openapi: 3.0.3
info:
  title: Policy Hub Public API v1
  description: |
    The Policy Hub Public API v1 is a comprehensive platform for managing and distributing API platform policies.
    It provides endpoints for discovering, versioning, and retrieving policy definitions, metadata, and documentation.
    Key features include policy search and filtering, version management, batch retrieval, and documentation access.
  version: 1.0.0
  contact:
    name: WSO2
    url: https://wso2.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.policyhub.wso2.com/api/v1
    description: Production server
...
paths:
  /policies:
    get:
      tags:
        - policies
      summary: List all policies
      operationId: listPolicies
      parameters:
        - name: search
          in: query
          description: Free text search over name, description, tags
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category (comma-separated for multiple)
          schema:
            type: string
        - name: categories
          in: query
          description: Filter by categories (comma-separated, alternative to category)
          schema:
            type: string
        - name: provider
          in: query
          description: Filter by provider (comma-separated for multiple)
          schema:
            type: string
        - name: providers
          in: query
          description: Filter by providers (comma-separated, alternative to provider)
          schema:
            type: string
        - name: platform
          in: query
          description: Filter by supported platform (comma-separated for multiple)
          schema:
            type: string
        - name: platforms
          in: query
          description: Filter by supported platforms (comma-separated, alternative to platform)
          schema:
            type: string
        - name: page
          in: query
          description: Page number
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          description: Number of items per page
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoliciesListResponse'
  /policies/resolve:
    post:
      tags:
        - policies
      summary: Retrieve multiple policies with strategy-based version selection
      description: |
        Resolve endpoint to retrieve multiple policies with strategy-based version selection.
        Supports exact version matching, latest patch/minor/major version resolution.
        Maximum 100 policies per request. Version format must be d.d.d (e.g., "1.2.3").
      operationId: resolvePolicies
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolvePolicyRequest'
      responses:
        '200':
          description: Resolve policy retrieval response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PoliciesWithDefinitionsResponse'
        '400':
          description: Invalid request payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /policies/{name}:
    get:
      tags:
        - policies
      summary: Get policy summary with latest version
      operationId: getPolicySummary
      parameters:
        - name: name
          in: path
          required: true
          description: Policy name
          schema:
            type: string
      responses:
        '200':
          description: Policy summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policies/{name}/versions:
    get:
      tags:
        - versions
      summary: List policy versions
      operationId: listPolicyVersions
      parameters:
        - name: name
          in: path
          required: true
          description: Policy name
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: List of policy versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyVersionsResponse'
  /policies/{name}/versions/latest:
    get:
      tags:
        - versions
      summary: Get latest policy version
      description: Returns the most recent version of a policy for easy UI navigation
      operationId: getLatestVersion
      parameters:
        - name: name
          in: path
          required: true
          description: Policy name
          schema:
            type: string
      responses:
        '200':
          description: Latest policy version detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '404':
          description: Policy not found or has no versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policies/{name}/versions/{version}:
    get:
      tags:
        - versions
      summary: Get policy version metadata
      operationId: getPolicyVersionDetail
      parameters:
        - name: name
          in: path
          required: true
          description: Policy name
          schema:
            type: string
        - name: version
          in: path
          required: true
          description: Policy version
          schema:
            type: string
      responses:
        '200':
          description: Policy version detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyResponse'
        '404':
          description: Policy version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policies/{name}/versions/{version}/definition:
    get:
      tags:
        - versions
      summary: Get raw policy definition (YAML format)
      operationId: getPolicyDefinition
      parameters:
        - name: name
          in: path
          required: true
          description: Policy name
          schema:
            type: string
        - name: version
          in: path
          required: true
          description: Policy version
          schema:
            type: string
      responses:
        '200':
          description: Raw policy definition
          content:
            text/yaml:
              schema:
                type: string
                description: Raw policy definition in YAML format
  /policies/{name}/versions/{version}/engine:
    get:
      tags:
        - versions
      summary: Get policy data for policy engine
      description: Returns policy metadata and definition optimized for policy engine consumption
      operationId: getPolicyForEngine
      parameters:
        - name: name
          in: path
          required: true
          description: Policy name
          schema:
            type: string
        - name: version
          in: path
          required: true
          description: Policy version
          schema:
            type: string
      responses:
        '200':
          description: Policy data for engine
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyWithDefinitionResponse'
        '404':
          description: Policy version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /policies/{name}/versions/{version}/docs:
    get:
      tags:
        - docs
      summary: Get all documentation pages
      operationId: getAllDocs
      parameters:
        - name: name
          in: path
          required: true
          description: Policy name
          schema:
            type: string
        - name: version
          in: path
          required: true
          description: Policy version
          schema:
            type: string
      responses:
        '200':
          description: All documentation pages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentationListResponse'
  /policies/{name}/versions/{version}/docs/{page}:
    get:
      tags:
        - docs
      summary: Get a single documentation page
      operationId: getSingleDoc
      parameters:
        - name: name
          in: path
          required: true
          description: Policy name
          schema:
            type: string
        - name: version
          in: path
          required: true
          description: Policy version
          schema:
            type: string
        - name: page
          in: path
          required: true
          description: Documentation page name
          schema:
            type: string
            enum: [overview, configuration, examples, faq]
      responses:
        '200':
          description: Documentation page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentationResponse'
        '404':
          description: Documentation page not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ResponseMeta:
      type: object
      properties:
        trace_id:
          type: string
          example: "abc123"
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          example: "xyz123"
      required:
        - trace_id
        - timestamp
    PaginatedResponseMeta:
      type: object
      properties:
        trace_id:
          type: string
          example: "abc123"
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          example: "xyz123"
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
      required:
        - trace_id
        - timestamp
        - pagination
    ErrorObject:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: POLICY_NOT_FOUND
        message:
          type: string
          description: Human-readable error message
          example: Policy not found
        details:
          type: object
          additionalProperties: true
          description: Optional structured context for debugging
      required:
        - code
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        data:
          nullable: true
          example: null
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      required:
        - success
        - meta
    PoliciesListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
        error:
          nullable: true
          example: null
        meta:
          $ref: '#/components/schemas/PaginatedResponseMeta'
      required:
        - success
        - meta
    PoliciesWithDefinitionsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/PolicyWithDefinition'
        error:
          nullable: true
          example: null
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      required:
        - success
        - meta
...
    PolicyResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Policy'
        error:
          nullable: true
          example: null
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      required:
        - success
        - meta
    PolicyVersionsResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Policy'
        error:
          nullable: true
          example: null
        meta:
          $ref: '#/components/schemas/PaginatedResponseMeta'
      required:
        - success
        - meta
    DocumentationListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/PolicyDocumentation'
        error:
          nullable: true
          example: null
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      required:
        - success
        - meta
    DocumentationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PolicyDocumentation'
        error:
          nullable: true
          example: null
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      required:
        - success
        - meta
    PolicyWithDefinitionResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/PolicyWithDefinition'
        error:
          nullable: true
          example: null
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      required:
        - success
        - meta
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        totalItems:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    Policy:
      type: object
      description: Standardized policy object used across all GET endpoints
      properties:
        name:
          type: string
          example: rate-limit
        version:
          type: string
          example: v1.1.0
        displayName:
          type: string
          example: Rate Limiting Policy
        description:
          type: string
          example: Rate limit policy with enhanced schema
        provider:
          type: string
          example: WSO2
        categories:
          type: array
          items:
            type: string
          example: ["security", "traffic-control"]
        tags:
          type: array
          items:
            type: string
          example: ["limit", "quota"]
        supportedPlatforms:
          type: array
          items:
            type: string
          example: ["apim-4.5+"]
        logoUrl:
          type: string
          format: uri
          example: https://raw.githubusercontent.com/wso2/policies/rate-limiting/1.1.0/logo.svg
        bannerUrl:
          type: string
          format: uri
          example: https://raw.githubusercontent.com/wso2/policies/rate-limit/v1.1.0/banner.png
        iconUrl:
          type: string
          format: uri
          example: https://raw.githubusercontent.com/wso2/policies/rate-limit/v1.1.0/icon.svg
        releaseDate:
          type: string
          format: date
          example: "2025-02-15"
        isLatest:
          type: boolean
          example: true
        sourceType:
          type: string
          example: github
        downloadUrl:
          type: string
          format: uri
          example: https://github.com/wso2/policies/rate-limit
      required:
        - name
        - version
        - displayName
        - provider
        - isLatest
    PolicyWithDefinition:
      type: object
      description: Streamlined policy object for engine/resolve operations including definition
      properties:
        name:
          type: string
          example: rate-limit
        version:
          type: string
          example: v1.1.0
        displayName:
          type: string
          example: Rate Limiting Policy
        provider:
          type: string
          example: WSO2
        categories:
          type: array
          items:
            type: string
          example: ["security", "traffic-control"]
        releaseDate:
          type: string
          format: date
          example: "2025-02-15"
        isLatest:
          type: boolean
          example: true
        sourceType:
          type: string
          example: github
        downloadUrl:
          type: string
          format: uri
          example: https://github.com/wso2/policies/rate-limit
        definition:
          type: string
          description: Raw policy definition in YAML format
          example: |
            policyName: rate-limit
            version: "1.1.0"
            schema:
              type: object
              properties:
                requestsPerMinute:
                  type: integer
                  minimum: 1
      required:
        - name
        - version
        - displayName
        - provider
        - categories
        - isLatest
        - definition
...
    ResolvePolicyRequest:
      type: object
      properties:
        policies:
          type: array
          items:
            $ref: '#/components/schemas/PolicyRequestItem'
          minItems: 1
          maxItems: 100
          example:
            - name: rate-limiting
              retrievalStrategy: exact
              baseVersion: "1.1.0"
            - name: jwt-authentication
              retrievalStrategy: latest_patch
              baseVersion: "2.1"
            - name: cors-policy
              retrievalStrategy: latest_minor
              baseVersion: "1"
            - name: api-throttling
              retrievalStrategy: latest_major
      required:
        - policies
    PolicyRequestItem:
      type: object
      properties:
        name:
          type: string
          example: rate-limiting
        retrievalStrategy:
          type: string
          enum: [exact, latest_patch, latest_minor, latest_major]
          example: exact
          description: |
            Strategy for version retrieval:
            - exact: Get the exact version specified in baseVersion
            - latest_patch: Get latest patch version within major.minor from baseVersion
            - latest_minor: Get latest minor version within major from baseVersion  
            - latest_major: Get latest major version (no baseVersion needed)
        baseVersion:
          type: string
          example: "1.1.0"
          description: |
            Base version for strategy-based retrieval. Required for exact, latest_patch, latest_minor.
            Format depends on strategy:
            - exact: "1.2.3" (full semantic version)
            - latest_patch: "1.2" (major.minor)
            - latest_minor: "1" (major only)
            - latest_major: not used
      required:
        - name
        - retrievalStrategy
...
    PolicyDocumentation:
      type: object
      properties:
        page:
          type: string
          enum: [overview, configuration, examples, faq]
        format:
          type: string
          enum: [markdown]
          example: markdown
        content:
          type: string
      required:
        - page
        - format
        - content